+++ common/src/main/java/tocraft/craftedcore/platform/PlatformData.java
@@ -2,8 +2,11 @@ package tocraft.craftedcore.platform;
 
 import dev.architectury.injectables.annotations.ExpectPlatform;
 import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.Nullable;
+import tocraft.craftedcore.config.Config;
 
 import java.lang.module.ModuleDescriptor.Version;
 import java.nio.file.Path;
@@ -41,6 +44,13 @@ public final class PlatformData {
     @ExpectPlatform
     @Contract(value = " -> !null", pure = true)
     public static ModLoader getModLoaderId() {
+        throw new AssertionError();
+    }
+
+    @ExpectPlatform
+    @ApiStatus.Internal
+    @Environment(EnvType.CLIENT)
+    public static void registerConfigScreen(String name) {
         throw new AssertionError();
     }